requirements.txt
asgiref==3.5.0
Django==3.2.15
django-filter==2.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
importlib-metadata==4.11.3
Markdown==3.3.7
Pillow==9.1.1
pytz==2021.1
requests==2.27.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
urllib3==1.26.9
zipp==3.8.0
coverage==6.4.4
flake8==3.8.4
mccabe==0.6.1
pycodestyle==2.6.0
pyflakes==2.2.0
